requirements.txt
torch
torchvision
ultralytics
Pillow
numpy
noise
psutil
requests
rich
tensorboard